pyproject.toml
[project]
name = "multi-document-rag-search-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "altair==6.0.0",
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.12.1",
    "attrs==25.4.0",
    "beautifulsoup4==4.14.3",
    "blinker==1.9.0",
    "cachetools==7.0.5",
    "certifi==2026.2.25",
    "charset-normalizer==3.4.5",
    "click==8.3.1",
    "colorama==0.4.6",
    "faiss-cpu==1.13.2",
    "filelock==3.25.1",
    "fsspec==2026.2.0",
    "gitdb==4.0.12",
    "gitpython==3.1.46",
    "h11==0.16.0",
    "hf-xet==1.3.2",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "huggingface-hub==1.6.0",
    "idna==3.11",
    "jinja2==3.1.6",
    "joblib==1.5.3",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jsonschema==4.26.0",
    "jsonschema-specifications==2025.9.1",
    "langchain==1.2.11",
    "langchain-core==1.2.18",
    "langgraph==1.1.0",
    "langgraph-checkpoint==4.0.1",
    "langgraph-prebuilt==1.0.8",
    "langgraph-sdk==0.3.11",
    "langsmith==0.7.16",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.3",
    "mdurl==0.1.2",
    "mpmath==1.3.0",
    "narwhals==2.18.0",
    "networkx==3.6.1",
    "numpy==2.4.3",
    "orjson==3.11.7",
    "ormsgpack==1.12.2",
    "packaging==26.0",
    "pandas==2.3.3",
    "pillow==12.1.1",
    "protobuf==6.33.5",
    "pyarrow==23.0.1",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pydeck==0.9.1",
    "pygments==2.19.2",
    "pypdf==6.8.0",
    "python-dateutil==2.9.0.post0",
    "pytz==2026.1.post1",
    "pyyaml==6.0.3",
    "referencing==0.37.0",
    "regex==2026.2.28",
    "requests==2.32.5",
    "requests-toolbelt==1.0.0",
    "rich==14.3.3",
    "rpds-py==0.30.0",
    "safetensors==0.7.0",
    "scikit-learn==1.8.0",
    "scipy==1.17.1",
    "sentence-transformers==5.2.3",
    "setuptools==82.0.1",
    "shellingham==1.5.4",
    "six==1.17.0",
    "smmap==5.0.3",
    "soupsieve==2.8.3",
    "streamlit==1.55.0",
    "sympy==1.14.0",
    "tavily-python==0.7.23",
    "tenacity==9.1.4",
    "threadpoolctl==3.6.0",
    "tiktoken==0.12.0",
    "tokenizers==0.22.2",
    "toml==0.10.2",
    "torch==2.10.0",
    "tornado==6.5.5",
    "tqdm==4.67.3",
    "transformers==5.3.0",
    "typer==0.24.1",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "tzdata==2025.3",
    "urllib3==2.6.3",
    "uuid-utils==0.14.1",
    "watchdog==6.0.0",
    "wikipedia==1.4.0",
    "xxhash==3.6.0",
    "zstandard==0.25.0",
]
